requirements.txt
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
Django==5.1.1
django-debug-toolbar==4.4.6
django-htmx==1.19.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
factory_boy==3.3.1
Faker==28.4.1
filelock==3.15.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
virtualenv==20.26.3
